+++ tests/dataflow/aie/test_vector.py
@@ -153,9 +153,64 @@ def _test_vector_vector_mul():
     print("Dataflow Simulator Passed!")
 
 
+def _test_vector_scalar_add_p0():
+    # https://github.com/Xilinx/mlir-aie/tree/main/programming_guide/section-2/section-2d
+    #                |--------------------------------------------|
+    #                v   v-------------------------v              v
+    # shim tile <-> mem tile <-> comp tile0    comp tile1    comp tile2
+    Ty = int32
+    M = 1024
+    P0 = 4
+    Mt = M // P0
+
+    @df.region()
+    def top():
+        @df.kernel(mapping=[P0])
+        def core(A: Ty[M], B: Ty[M]):
+            pi = df.get_pid()
+            B[pi * Mt : (pi + 1) * Mt] = allo.add(A[pi * Mt : (pi + 1) * Mt], 1)
+
+    mod = df.build(top, target="aie")
+    A = np.random.randint(0, 100, M).astype(np.int32)
+    B = np.zeros(M).astype(np.int32)
+    mod(A, B)
+    np.testing.assert_allclose(B, A + 1)
+    print("PASSED!")
+
+
+def _test_vector_vector_add_p0():
+    #                  |--------------------------------------------|
+    #                  v   v--------------------------v             v
+    # shim tile <-> A mem tile 0 <-> comp tile0    comp tile1    comp tile2
+    #       ^-----> B mem tile 1 <-------^------------^-------------^
+    Ty = int32
+    M = 1024
+    P0 = 4
+    Mt = M // P0
+
+    @df.region()
+    def top():
+        @df.kernel(mapping=[P0])
+        def core(A: Ty[M], B: Ty[M], C: Ty[M]):
+            pi = df.get_pid()
+            C[pi * Mt : (pi + 1) * Mt] = allo.add(
+                A[pi * Mt : (pi + 1) * Mt], B[pi * Mt : (pi + 1) * Mt]
+            )
+
+    mod = df.build(top, target="aie")
+    A = np.random.randint(0, 100, M).astype(np.int32)
+    B = np.random.randint(0, 100, M).astype(np.int32)
+    C = np.zeros(M).astype(np.int32)
+    mod(A, B, C)
+    np.testing.assert_allclose(C, A + B)
+    print("PASSED!")
+
+
 if __name__ == "__main__":
     _test_vector_scalar_add()
     _test_vector_scalar_mul()
     _test_vector_vector_add()
     _test_vector_vector_bf16_add()
     _test_vector_vector_mul()
+    _test_vector_scalar_add_p0()
+    _test_vector_vector_add_p0()
